Added non existing debt id check
Added if statement to check if the debt exist in the DB and change error type from return.

+++ HardShipAPI/Models/Debt.cs
@@ -1,6 +1,6 @@
 ﻿public class Debt
 {
-    public int DebtID { get; set; }
+    public long DebtID { get; set; }
     public string Name { get; set; }
     public string DOB { get; set; }
     public decimal? Income { get; set; }
