+++ HardShipAPI/Models/HardshipManagement.cs
@@ -3,7 +3,7 @@
     public class HardshipManagementInsert
     {
         public short HardshipTypeID { get; set; }
-        public int DebtID { get; set; }
+        public long DebtID { get; set; }
         public string Name { get; set; }
         public string DOB { get; set; }
         public decimal Income { get; set; }
@@ -25,8 +25,8 @@
     public class HardshipManagementView
     {
         public string HardshipTypeName { get; set; }
-        public int HardshipID { get; set; }
-        public int DebtID { get; set; }
+        public long HardshipID { get; set; }
+        public long DebtID { get; set; }
         public string Name { get; set; }
         public string DOB { get; set; }
         public decimal? Income { get; set; }
